agregando las Interfaces PacienteReposity y TurnoRepository y ordenando el codigo en las capas Controller y varios

+++ src/main/java/com/backend/clinicaDental/service/IPacienteService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface IPacienteService {
     PacienteSalidaDto registrarPaciente(PacienteEntradaDto paciente);
     List<PacienteSalidaDto> listarPacientes();
-    PacienteSalidaDto buscarPacientePorId(int id);
+    PacienteSalidaDto buscarPacientePorId(Long id);
 }
